refactor(工具输出): 展平内容存储契约
