+++ services/wisepen-chat-service/src/chat/application/tools/common/tool_content_store/__init__.py
@@ -1,11 +1,11 @@
-from .core import (
+from .models import (
     StoredToolContent,
     ToolContentChunk,
     ToolContentIndex,
     ToolContentIndexEntry,
     ToolContentReceipt,
-    ToolContentRepository,
 )
+from .repository import ToolContentRepository
 from .store import ToolContentPutResult, ToolContentPutStatus, ToolContentStore
 
 __all__ = [
