--- services/wisepen-chat-service/src/chat/application/tools/common/tool_content_store/core/__init__.py
@@ -1,17 +1,0 @@
-from .models import (
-    StoredToolContent,
-    ToolContentChunk,
-    ToolContentIndex,
-    ToolContentIndexEntry,
-    ToolContentReceipt,
-)
-from .repository_protocol import ToolContentRepository
-
-__all__ = [
-    "StoredToolContent",
-    "ToolContentChunk",
-    "ToolContentIndex",
-    "ToolContentIndexEntry",
-    "ToolContentReceipt",
-    "ToolContentRepository",
-]
