+++ services/wisepen-chat-service/src/chat/application/tools/common/tool_content_store/models.py
